secp256k1: Unexport idents that do not need to be.
This unexports various functions, structs, and constants that were
exported for some reason when porting this from the upstream project
when they really should not have been since they deal with package
internals.

While here, correct a few other minor differences that were introduced
between the package and the upstream code and correct a few comments
that appear to have had a blanket find/replace applied to them
incorrectly.