secp256k1: Improve NonceRFC6979 comment.
The comment indicated it was to generate an ECDSA nonce, but that is not
the case.  Instead, it is a method to generate a deterministic nonce
based on a private key and message which for general use in signing
applications.  As a case in point, it is used with Schnorr signing as
well.

Also, call out the semantics of the additional optional fields while
here.