secp256k1: Move group operations to new curve.go.
Now that the group point addition and point doubling funcs have been
made stadalone, move them into a new file named curve.go which is
intended to ultimately house all of the ec operations.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.