secp256k1: Make k splitting func standalone.
This makes the function designed to decompose a scalar value into its
balanced length-two representation in order take advantage of the
endomorphism when performing scalar multiplication standalone instead of
being defined on the curve used to satisfy the crypto/elliptic stdlib
interface.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.