secp256k1: Make k mod reduce func standalone.
This makes the function designed to reduce a scalar value modulo the
curve order standalone instead of being defined on the curve used to
satisfy the crypto/elliptic stdlib interface.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.