secp256k1: Refactor scalar mult logic from adaptor.
This refactors the functions dealing with scalar multiplication and the
related affine conversions out of the adpator code into the local
curve.go file and updates most of the callers to use the internal
version instead.

The API could really use a cleanup, but it has been kept consistent for
now.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.