secp256k1: Add optimized sqrt field calc.
This introduces the ability to calculate the square root modulo the
field characteristic, when it exists, via the specialized field value
type.  It is significantly faster than performing the calculation via
big integers and it is also constant time.

The motivation for adding this is that it is needed for point
decompression which currently makes use of big ints.

Comprehensive tests with 100% coverage and a full suite of benchmarks
are included.

Like other code in this area, the code relies on techniques that are not
the easiest to review for correctness, so significant effort has been
put into attempting to clearly explain the details via comments to help
with review.

For now, this merely introduces the method without modifying any of the
code to make use of it.  The intention is to do so in future commits.

The following benchmark shows a comparison between calculating the
square root via the new specialized method and via generic big ints.
As can be seen, the new method is roughly twice as fast and  reduces
allocations by 100%.

benchmark       old ns/op    new ns/op    delta
--------------------------------------------------
BenchmarkSqrt   24605        11087        -54.94%

benchmark       old allocs   new allocs   delta
--------------------------------------------------
BenchmarkSqrt   91           0            -100.00%

benchmark       old bytes    new bytes    delta
--------------------------------------------------
BenchmarkSqrt   1016         0            -100.00%