secp256k1: Add field func to determine when >= P-N.
This introduces the ability to determine if a specialized field value is
greater than or equal to the field prime minus the group order in
constant time.  It is also significantly faster than performing the
calculation via big integers since it would require a conversion from
the specialized type in practice.

Comprehensive tests with 100% coverage and benchmarks are included.

For now, this merely introduces the method without modifying any of the
code to make use of it, but it will be useful when working with
signatures and public key recovery in the future.

The following benchmark shows a comparison between determining the
condition via the new specialized method and via generic big ints:

benchmark                          old ns/op    new ns/op    delta
--------------------------------------------------------------------
BenchmarkIsGtOrEqPrimeMinusOrder   123          10.6         -91.38%

benchmark                          old allocs   new allocs   delta
--------------------------------------------------------------------
BenchmarkIsGtOrEqPrimeMinusOrder   2            0            -100.00%

benchmark                          old bytes    new bytes    delta
--------------------------------------------------------------------
BenchmarkIsGtOrEqPrimeMinusOrder   96           0            -100.00%