secp256k1: Return num instead of bool for overflow.
This modifies the SetBytes method on the ModNScalar type to return the
overflow condition as a uint32 instead of a bool in order to better
facilite constant time code in upcoming changes and updates the callers
as needed.

It also updates the internal fieldToModNScalar convenience function to
return the flag as well.