secp256k1: Rework DER signature parsing code.
This reworks the DER signature parsing code to use the majority of the
implementation code that exists in the consensus scripting engine,
updates it to make use of the new specialized mod N scalar, and
introduces a new error type and associated error code for signature
errors which can be used with the errors.Is and errors.As interfaces to
programmatically detect the failure reason.

It includes full tests for the new signature error infrastructure.

The primary motivation behind using the implementation code from the
consensus script engine is that it is already extremely well tested, any
signatures that don't conform to it would be rejected by consensus
anyway, and the code is easier to follow due to the use of constants and
better documentation.