secp256k1: Export ScalarBaseMultNonConst.
This exports the scalarBaseMultJacobian func as ScalarBaseMultNonConst
in order to allow external callers to perform optimized elliptic curve
point multiplication of the group base point with a scalar in Jacobian
project coordinates in addition to more clearly indentifying that the
function is not constant time.

This is part of exposing the ec group operations so callers are not
forced to round trip through big ints in affine space.