secp256k1: Rework pubkey tests.
This reworks the tests related to public keys to them more consistent
with the rest of the package and improve their coverage to include all
of the possible error paths.

It also brings back the hybrid pubkey parsing to match what the function
comment claims and corrects the tests accordingly.  The restriction as
to which specific types of pubkeys are allowed must be (and is) enforced
by the opcode(s) that interpret them as opposed to the lib that
implements the more generic specification.

The following is an overview of the changes:

- Convert all tests to use hex strings
- Add and verify the expected x and y coordinates from a parsed pubkey
- Bring back support for parsing hybrid keys
- Correct remaining existing tests for hybrid pubkeys
- Add negative tests for all combinations of encoding what would
  otherwise be valid pubkey with the wrong format byte and y coord
  oddness
- Add negative tests for x and y coordinates == p for the uncompressed,
  compressed, and hybrid formats
- Add negative tests for all formats such that either the x or y
  coordinate are > p and are constructed in such a way that they would
  be valid if the coordinate is taken mod p to ensure implementations
  don't incorrectly ignore the range check