secp256k1: Optimize precomp values to use affine.
This optimizes the pre-computed byte points used to accelerate scalar
base multiplication to store the data in affine coordinates instead of
Jacobian coordinates which reduces the memory usage requirement to 66%
of what it current requires and also has the important benefit of
further speeding up the computation.

This is the case because projecting affine coordinates into Jacobian
space is essentially free and the point doubling and addition routines
have optimizations which allow them to avoid additional operations when
the Z coordinate is 1, which is the case for an initial affine
projection.

Further, since the compressed table is stored in the string table of the
binary, it also reduces the size the of final binary by ~385KiB.

The following benchmark shows a before and after comparison of scalar
base multiplication as well as how that translates to signature
verification:

name                       old time/op    new time/op    delta
-------------------------------------------------------------------------
ScalarBaseMult        34.5µs ± 1%   24.7µs ± 1%   -28.43% (p=0.008 n=5+5)
ScalarBaseMultLarge   48.2µs ± 1%   38.0µs ± 1%   -21.08% (p=0.008 n=5+5)
SigVerify              181µs ± 5%    163µs ± 2%    -9.86% (p=0.008 n=5+5)

While 18 µs less per signature verification might not seem like much on
the surface, consider that every transaction requires at least one
signature operation, so there are a ton of them when doing no checkpoint
syncs.  For a concrete number, verifying 100 million signatures would
take 30 minutes less time.